+++ components/tutors-ui/lib/Molecules/Cards/UnitCard.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
-    import VideoCard from "../cards/VideoCard.svelte";
-    import TalkCard from "../cards/TalkCard.svelte";
-    import { CardDeck, Image, NoteCard } from "tutors-ui";
+    import { CardDeck, Image, NoteCard, TalkCard, VideoCard } from "tutors-ui";
     import type { Lo } from "tutors-reader-lib/src/types/lo-types";
-    import { layout, currentCourse } from "../../stores";
+    import { layout, currentCourse } from "tutors-course-reader/src/stores";
     import { onDestroy } from "svelte";
     export let unit: Lo;
     const panelVideos = unit.los.filter((lo) => lo.type == "panelvideo");
